+++ tests/test_predict.py
@@ -1,4 +1,5 @@
 import gym
+import numpy as np
 import pytest
 import torch as th
 
@@ -14,6 +15,24 @@ MODEL_LIST = [
     SAC,
     DQN,
 ]
+
+
+class SubClassedBox(gym.spaces.Box):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+
+class CustomSubClassedSpaceEnv(gym.Env):
+    def __init__(self):
+        super().__init__()
+        self.observation_space = SubClassedBox(-1, 1, shape=(2,), dtype=np.float32)
+        self.action_space = SubClassedBox(-1, 1, shape=(2,), dtype=np.float32)
+
+    def reset(self):
+        return self.observation_space.sample()
+
+    def step(self, action):
+        return self.observation_space.sample(), 0.0, np.random.rand() > 0.5, {}
 
 
 @pytest.mark.parametrize("model_class", MODEL_LIST)
@@ -80,3 +99,12 @@ def test_dqn_epsilon_greedy():
     # is vectorized should not crash with discrete obs
     action, _ = model.predict(obs, deterministic=False)
     assert env.action_space.contains(action)
+
+
+@pytest.mark.parametrize("model_class", [A2C, SAC, PPO, TD3])
+def test_subclassed_space_env(model_class):
+    env = CustomSubClassedSpaceEnv()
+    model = model_class("MlpPolicy", env, policy_kwargs=dict(net_arch=[32]))
+    model.learn(300)
+    obs = env.reset()
+    env.step(model.predict(obs))